requirements.txt
flask
nltk
pandas
Pillow
scikit-learn
